got rid of repeated delete(gca) statements
stringing multiple delete(gca) statements together is a shotgun
approach to clearing a figure.  It has no idea what is in the layer.
The code isn’t clear as to which axes are supposed to be deleted.
Instead, we’re deleting all the axes children from the chosen figure.
using :
findobj(fignum,’Type’,’axes’)
Also, the way matlab works, multiple plots can be in the same axes, so
this isn’t really necessary anyway.