consolidating the many dialog boxes that ask for both grid parameters and event selection parameters
Every one of these dialog boxes is a variation on a theme with no
coherency at all.  I’ve added two new classes are used to provide
modularity and a single-place where they all can be changed.

The classes are GridParameterChoice (chosing ni, ra) and
EventSelectionChoice (chosing dx,dy,dz, load, and save).  The dialog
boxes still have additional parameters floating in odd locations, but
the parts that work have much improved usability.

Each of these dialog boxes are the gateways to some mapping routine
that requires a grid.  The grid creation should also be consolidated,
but this will happen after the interfaces are fixed.

bvalgrid is the routine that has the most work done on it.