created a class for XY grids
ZmapGrid.m can be used to create grids that are used for calculations
and map displays throughout ZMAP.  it requires only enough details to
create X and Y vectors, which are then turned into grids.  It can also
filter these based on whether they are contained within a polygon.
Either the raw or filtered grids can be plotted. Plots are reused where
possible, and are tied to the name that the grid is given.

This should clean up function code greatly.