shapes shapes shapes
The Grid issues led to the circle and polygon issues.  There are so
many ways to create a circle, and they are used in many places.  I now
have one dialog box that could work everywhere. it’s attached to a menu
(that should work everywhere).  I only need to hook up the radio
buttons to the behaviors of nearest/radius/both. and make sure it gets
plotted when selected.  THEN I can move on and start implementing the
views again.