implemented gridding and the quarry function
The general theme lately is that of getting the gridding functions to
work better.  A typical function requires a grid over which a function
will be calculated.  The grid is generally defined via user-interactive
functions, like GridParameterChoice, which can be used to get a
ZmapGrid.  The method for selecting events at each grid point is
controlled by the user via EventSelectionChoice.  This allows them to
choose between #events nearby and radius.   Both of these options can
be used by ZmapFunctionDlg.m, which greatly simplifies the creation of
function dialog boxes.
With a grid and a method of event selection in hand, then gridfun.m can
be used (which behaves somewhat like cellfun, structfun, etc).  This
takes a function handle, and uses it to calculate a value at each grid
point using the appropriate selection method.  The results can be shown
with the gridpcolor function, which compensates for matlab’s pcolor
function.

In the original pcolor, if you pass a 3x3 grid, with 3 x’s and 3 y’s,
then you’ll end up with a 2x2 box where the values are the grid
intersection points. This is definitely not what we want.  So
gridpcolor takes the same parameters, and treats each point as the
CENTER of the grid.

Other notable changes:

fixed problem where functions looked for the main_map_ax when they
should look for mainmap_ax

modified the ZmapCatalog’s selectRadius to return the max distance, so
it works like the selectClosestEvents.

allowed a ZmapGrid to be created from the structure generated from
GridParameterChoice.

selgp updated, though it’s functionality can now be handled by the
ZmapGrid.

fix_caxis.m : consolidated a bunch of fix_axis functions together into
one function.

findquar.m : now uses the ZmapFunction paradigm to do its job.  Other
plotting functions will follow suit.