introduced ZmapCatalogView
this could be a great solution for the trouble with polygon behavior.
It will also make plotting slices easier.
ZmapCatalogView is a window onto one of the global catalogs, but
filtered.. date,mag, location, polygon, etc. If the view is plotted
using linkedplot, then whenever the view changes, the plot will
automatically change too!