rewriting ZmapDialog
this is to make it easier to add other items. the old one was
hard-wired. now I can do radio buttons, and add other items on the fly.

Currently, however, the ability to disable features using a checkbox is
disabled.  I may need to write the ability for a ZmapDialogPart to
report back the handles for things that can be deleted.