Making grids and shapes lat/lon aware
the ZmapGrid no longer assumes that dx is constant with dy.  Both X and
Y values are contained in matrices now, instead of relying upon
vectors.
Shapes now have ability to scale & translate linearly, or with lat/lon
awareness. This means a circle can still remain a circle of the same
size whether it is moved around from  equator to high latitudes.