ZMAPGRID now should work with matrices
Must be tested, though.  epecially gridfun
One behavioral change: if shape is requested, but none exists, it gives
a warning, but doesn’t ask you to select a new shape.
The ZmapGrid is capable of handlng 3-d matrices, but that isn’t
implemented in the constructor yet.
To do: modify create_grid to leverage this.Make sure rest of code base
can handle this.