added smart figure creation
if a figure gets its position via `position_in_current_monitor` , then
it will be placed on the active monitor (whichever monitor the mouse is
in).  If figure is too tall, and a taller monitor is available, it will
be placed there instead.  One can specify center, left, or right, along
with a border.