performance fixes and shape-xsection fixes
Testing with a huge dataset inspired by a talk with Paul S., realized a
few big places that could change.  Added a smarter distance2curve,
which greatly reduced time of calculating xsections.
Cross sections now are filtered by the active shape.
Moment calculating is more forgiving if not provided a catalog
if there are lots of events, the marker is changed to make for faster
drawing.