added scatter plots that could have arbitrary axes, color, size
CatalogExploration plot generates a scatter plot with interactive axes.
Clicking on the axes or the data will provide an opportunity to change
that axes, or the color, or the size.
This has been implemented at the cross-section level.  I didn’t
implement it for the main map because that is, by definition, a main
map.
The cross section display acts funky when a shape is involved. I’ll
have to investigate that