wandering around in the misfit section
I was originally looking for another function to implment, but got
sucked down the misfit hole.  I consolidated it somewhat, and made some
sense of magic global values (mi, for example). This took it a step
closer to being usable.  But now I’ll put aside and look into the next
item