changed color-by values
now can color by depth, date, or magnitude.
colorbar has been modifed (as has the colorbar button) to intercept and
interpret date values, flip depths, and give a label to it.

tried to also reign in the message boxes, [msgbox_nobutton.m] but
realize i don’t have control of the position of the input box.