BROKEN STAGE: grid updating
updating ZmapGrid and friends to handle ellipsoids.  This allows multiscale catalogs (micron [matlab's term, not mine], through km], and as a side-effect could also enable this to be used on other bodies.. eg. mars.  At this stage, ZmapGrid is essentially updated, but ZmapGrid constructors throughout the program should also be updated.