downloading from fdsn just became smarter
for requests that exceed max # of downloadable events, it will break it into pieces and then download.  This depends, however, on the error message provided by the service, and is therefore somewhat fragile.

I have done away with the parallel processing and plotting during downloads, as it was complex and buggy. It has been replaced with a status window that appears and counts the number of bytes received. So, after the service starts returning data, you will know that something is happening.