Fixing colorbar
cleaned up zmap starting by putting the critical "CoordinateSYstem" into apdata, where everyone can find it, without having to create a funny constructor for ZmapData.

Colorbar is fixed.  It didn't work once this was set to xyz. But now the colorable fields live in the catalog class.
Dates show properly in th ecolorbar now.
For main map, the colorbar will always reappear if you have the data set to color-by something, and disappear if ypu set it to "-none-".