Grid chooser fix
plotting area more accurately reflects main map axes. Upon opening, the limits and aspect ratio are set to match the main map. Also, if a shape is provided, then that shape is also drawn.

Changes to the main map while the grid_chooser window is open are not reflected in the grid_chooser window until it is closed and reopened.