moved save directory outside of zmap
worked on some of the cross section methods
fixed bug on loading preferences  app
zmap no longer crashes when no catalog present on restart
added a method (that may change) to cut catalog and crop based on current selections
Fixed FDSN import when request is too large. It treated the max allowable, and max retrieved as the same thing.